removed api logging
